pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alexandre Klostermann <alexandre@aupi.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = "^0.30.11"
crewai-tools = "^0.2.6"
langchain-google-genai = "1.0.3"
python-dotenv = "^1.0.1"
django = "^4.2.2"
prettyconf = "^2.2.1"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
django-rest-passwordreset = "^1.4.1"
django-cors-headers = "^4.4.0"
django-colorfield = "^0.11.0"
stripe = "^10.12.0"
django-username-email = "^2.5.6"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
